+++ src/registers/commands/installEap.ts
@@ -15,68 +15,73 @@ import * as path from 'path'
 import buildEap from '../../generate/eapBuild'
 import { EdgerosTreeItem } from '../../components/treeItem'
 import { uploadEap, installEap } from '../../utility/edgerosApi'
-import { getGlobalState, getWorkspaceSettings, selectBuildPath } from '../../common'
+import { selectedDevice, getWorkspaceSettings, selectBuildPath } from '../../common'
+import { assert } from '../../errors'
 
 export = function (context: vscode.ExtensionContext) {
-  const disposable = vscode.commands.registerCommand('edgeros.installEap', async (...options: EdgerosTreeItem[]) => {
+  context.subscriptions.push(vscode.commands.registerCommand('edgeros.installEap',
+    (...options: EdgerosTreeItem[]) => {
+      return install(context, options)
+    })
+  )
+
+  context.subscriptions.push(vscode.commands.registerCommand('edgeros.installFromEapFile',
+    (...options: EdgerosTreeItem[]) => {
+      return install(context, options, true)
+    })
+  )
+};
+
+async function install (context: vscode.ExtensionContext, options: EdgerosTreeItem[], selectFile = false) {
+  try {
+    const workspaceSettings = getWorkspaceSettings()
+
     try {
-      const workspaceSettings = getWorkspaceSettings()
-      let workspaceFoldersExit = false
-      if (vscode.workspace.workspaceFolders) workspaceFoldersExit = true
+      const insideWorkspace = vscode.workspace.workspaceFolders
+      const devInfo = assert(await selectedDevice(context, options), 'No EdgerOS device is selected')
 
-      try {
-        // get egeros config
-
-        const devList = getGlobalState(context)
-        const devInfo = devList?.find(item => {
-          return item.devName === options[0].label
+      let eapPath: string
+      // 弹出选择框
+      if (selectFile || workspaceSettings.installEAP === 'Manual' || !insideWorkspace) {
+        eapPath = await selectEapFile()
+      } else if (workspaceSettings.installEAP === 'Auto') {
+        const projectDir = await selectBuildPath(context)
+        eapPath = await buildEap(projectDir, {
+          configInfo: workspaceSettings
         })
-
-        if (!devInfo) {
-          return
-        }
-
-        let eapPath: string
-        // 弹出选择框
-        if (workspaceSettings.installEAP === 'Manual' || workspaceFoldersExit === false) {
-          const eapNames: vscode.Uri[] | undefined = await vscode.window.showOpenDialog({
-            canSelectMany: false,
-            filters: {
-              // eslint-disable-next-line @typescript-eslint/naming-convention
-              'EdgerOS App': ['zip', 'eap']
-            }
-          })
-          if (eapNames) {
-            eapPath = eapNames[0].fsPath
-          } else {
-            throw new Error('File selection cancelled')
-          }
-        } else if (workspaceSettings.installEAP === 'Auto') {
-          const projectDir = await selectBuildPath(context)
-          eapPath = await buildEap(projectDir, {
-            configInfo: workspaceSettings
-          })
-        }
-
-        // Progress 动效
-        await vscode.window.withProgress({
-          location: vscode.ProgressLocation.Window,
-          title: 'EdgerOS',
-          cancellable: false
-        }, async (progress, token) => {
-          progress.report({ message: 'Upload EdgerOS App' })
-          await uploadEap(eapPath, devInfo.devIp, devInfo.devPwd)
-          progress.report({ message: 'Install EdgerOS App' })
-          const installMsg = await installEap(eapPath.split(path.sep).pop() as string, devInfo.devIp, devInfo.devPwd)
-          return installMsg
-        })
-        vscode.window.showInformationMessage('Install app success')
-      } catch (err) {
-        vscode.window.showErrorMessage('Install EdgerOS App : ' + err)
       }
-    } catch (err: any) {
-      vscode.window.showErrorMessage(err.message)
+
+      // Progress 动效
+      await vscode.window.withProgress({
+        location: vscode.ProgressLocation.Window,
+        title: 'EdgerOS',
+        cancellable: false
+      }, async (progress, token) => {
+        progress.report({ message: 'Upload EdgerOS App' })
+        await uploadEap(eapPath, devInfo.devIp, devInfo.devPwd)
+        progress.report({ message: 'Install EdgerOS App' })
+        const installMsg = await installEap(eapPath.split(path.sep).pop() as string, devInfo.devIp, devInfo.devPwd)
+        return installMsg
+      })
+
+      vscode.window.showInformationMessage('Install app success')
+    } catch (err) {
+      vscode.window.showErrorMessage('Install EdgerOS App : ' + err)
+    }
+  } catch (err: any) {
+    vscode.window.showErrorMessage(err.message)
+  }
+}
+
+async function selectEapFile () {
+  const eapNames = await vscode.window.showOpenDialog({
+    canSelectMany: false,
+    filters: { // eslint-disable-next-line @typescript-eslint/naming-convention
+      'EdgerOS App': ['zip', 'eap']
     }
   })
-  context.subscriptions.push(disposable)
-};
+  if (!eapNames) {
+    throw new Error('File selection cancelled')
+  }
+  return eapNames[0].fsPath
+}