+++ src/registers/commands/openConsole.ts
@@ -5,22 +5,13 @@
  * @Last Modified time: 2021-04-16 17:12:57
  */
 import * as vscode from 'vscode'
-import { getGlobalState } from '../../common'
+import { selectedDevice } from '../../common'
 import { EdgerosTreeItem } from '../../components/treeItem'
 import { connect } from '../../components/console'
 
 export = function (context: vscode.ExtensionContext) {
-  const disposable = vscode.commands.registerCommand('edgeros.openConsole', (...options: EdgerosTreeItem[]) => {
-    if (options && options.length > 0) {
-      const templastDevice = options[0]
-      const deviceList = getGlobalState(context)
-      const devData = deviceList?.find(item => {
-        return item.devName === templastDevice.label
-      })
-      connect(devData)
-    } else {
-      connect()
-    }
+  const disposable = vscode.commands.registerCommand('edgeros.openConsole', async (...options: EdgerosTreeItem[]) => {
+    connect(await selectedDevice(context, options))
   })
   context.subscriptions.push(disposable)
 };